Interval Change
In order to build a contrasting experimental environment, a simulation program with a node spacing of 6.5~8.5m and a packet sending interval of 0.00015~0.00045s was designed.
The change period of the node distance is T (40s), and the period of sending interval is T/4.

In the future, we will add Payload Size change, which change period is T/2.